refactor: /academics api 폴더구조 수정
