+++ apis/v2/academics/courses/[code].ts
@@ -1,0 +1,4 @@
+import { deleteRequest } from '@/apis';
+
+export const deleteCourse = async (code: string) =>
+  deleteRequest(`/v2/academics/courses/${code}`, { jsessionID: true });
