+++ app/[locale]/academics/graduate/course-changes/page.tsx
@@ -1,6 +1,6 @@
 import { Metadata } from 'next';
 
-import { getCourseChanges } from '@/apis/academics';
+import { getCourseChanges } from '@/apis/v1/academics/[type]/course-changes';
 import { getMetadata } from '@/utils/metadata';
 import { graduateCourseChanges } from '@/utils/segmentNode';
