+++ app/[locale]/academics/graduate/courses/page.tsx
@@ -1,4 +1,4 @@
-import { getCourses } from '@/apis/academics';
+import { getCourses } from '@/apis/v2/academics/courses';
 import PageLayout from '@/components/layout/pageLayout/PageLayout';
 import { Language } from '@/types/language';
 import { getMetadata } from '@/utils/metadata';
