+++ app/[locale]/academics/graduate/scholarship/[id]/page.tsx
@@ -1,4 +1,4 @@
-import { getScholarship } from '@/apis/academics';
+import { getScholarship } from '@/apis/v2/academics/scholarship/[id]';
 import InvalidIDFallback from '@/components/common/InvalidIDFallback';
 import { Language } from '@/types/language';
 import { getMetadata } from '@/utils/metadata';
