+++ app/[locale]/academics/undergraduate/course-changes/edit/page.tsx
@@ -1,4 +1,4 @@
-import { getCourseChanges } from '@/apis/academics';
+import { getCourseChanges } from '@/apis/v1/academics/[type]/course-changes';
 import { getPath } from '@/utils/page';
 import { undergraduateCourseChanges } from '@/utils/segmentNode';
