+++ app/[locale]/academics/undergraduate/course-changes/page.tsx
@@ -1,4 +1,4 @@
-import { getCourseChanges } from '@/apis/academics';
+import { getCourseChanges } from '@/apis/v1/academics/[type]/course-changes';
 import { getMetadata } from '@/utils/metadata';
 import { undergraduateCourseChanges } from '@/utils/segmentNode';
