+++ app/[locale]/academics/undergraduate/courses/page.tsx
@@ -1,4 +1,4 @@
-import { getCourses } from '@/apis/academics';
+import { getCourses } from '@/apis/v2/academics/courses';
 import PageLayout from '@/components/layout/pageLayout/PageLayout';
 import { Course } from '@/types/academics';
 import { Language } from '@/types/language';
