+++ app/[locale]/admissions/international/scholarships/page.tsx
@@ -1,6 +1,6 @@
 import { Metadata } from 'next';
 
-import { getInternationalScholarships } from '@/apis/admission';
+import { getInternationalScholarships } from '@/apis/v1/admissions/international/scholarships';
 import HTMLViewer from '@/components/editor/HTMLViewer';
 import PageLayout from '@/components/layout/pageLayout/PageLayout';
 import { getMetadata } from '@/utils/metadata';
